+++ packages/in_app_purchases_paywall_ui/lib/paywall/moritz/moritz_restore_row.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:in_app_purchases_intl/helper/paywall_helper.dart';
+import 'package:in_app_purchases_paywall_ui/paywall/inherit/paywall_data_iw.dart';
 import 'package:in_app_purchases_paywall_ui/paywall/inherit/subscription_callback_iw.dart';
 
 class MoritzRestoreRow extends StatelessWidget {
@@ -10,6 +12,9 @@ class MoritzRestoreRow extends StatelessWidget {
         onPressed: () {
           subscriptionData.onRestoreTap();
         },
-        child: Text("Restore", style: Theme.of(context).textTheme.button));
+        child: Text(
+            PaywallDataIW.of(context).restoreText ??
+                PaywallL10NHelper.of(context).restore_purchase,
+            style: Theme.of(context).textTheme.button));
   }
 }