+++ packages/in_app_purchases_paywall_ui/lib/paywall/simple/legal_row.dart
@@ -1,73 +1,62 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-import 'package:in_app_purchases_paywall_ui/in_app_purchases_paywall_ui.dart';
+import 'package:in_app_purchases_paywall_ui/paywall/inherit/paywall_data_iw.dart';
 
 /// Simple Legal Row with TOS and PP
 class LegalRow extends StatelessWidget {
-  final ChromeSafariBrowser browser = new ChromeSafariBrowser();
-  final ThemeData theme;
-  final TextAndUrl? tosData;
-  final TextAndUrl? ppData;
-
-  LegalRow(this.theme, this.tosData, this.ppData);
-
-  /// Opens Browser onTap with the given URL
-  _onTapTos() async {
-    await browser.open(
-        url: Uri.parse(tosData!.url),
-        options: ChromeSafariBrowserClassOptions(
-            android: AndroidChromeCustomTabsOptions(),
-            ios: IOSSafariOptions(barCollapsingEnabled: true)));
-  }
-
-  /// Opens Browser onTap with the given URL
-  _onTapPp() async {
-    await browser.open(
-        url: Uri.parse(ppData!.url),
-        options: ChromeSafariBrowserClassOptions(
-            android: AndroidChromeCustomTabsOptions(),
-            ios: IOSSafariOptions(barCollapsingEnabled: true)));
-  }
-
-  Widget get tosItem => tosData == null
-      ? Container()
-      : GestureDetector(
-          onTap: _onTapTos,
-          child: Text(
-            tosData?.name ?? "Terms of Service",
-            style: TextStyle(
-                decoration: TextDecoration.underline,
-                color: theme.textTheme.button?.color ?? theme.primaryColor,
-                fontSize: 12),
-          ),
-        );
-
-  Widget get ppItem => ppData == null
-      ? Container()
-      : GestureDetector(
-          onTap: _onTapPp,
-          child: Text(
-            ppData?.name ?? "Privacy Policy",
-            style: TextStyle(
-                decoration: TextDecoration.underline,
-                color: theme.textTheme.button?.color ?? theme.primaryColor,
-                fontSize: 12),
-          ),
-        );
+  const LegalRow({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final tosData = PaywallDataIW.of(context).tosData;
+    final ppData = PaywallDataIW.of(context).ppData;
     return Row(
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
-        Container(
-          padding: EdgeInsets.only(left: 2, right: 2),
-          child: tosItem,
-        ),
-        Container(
-          padding: EdgeInsets.only(left: 2, right: 2),
-          child: ppItem,
-        )
+        if (tosData != null)
+          Container(
+            padding: EdgeInsets.only(left: 2, right: 2),
+            child: GestureDetector(
+              onTap: () async {
+                final ChromeSafariBrowser browser = new ChromeSafariBrowser();
+                await browser.open(
+                    url: Uri.parse(PaywallDataIW.of(context).tosData!.url),
+                    options: ChromeSafariBrowserClassOptions(
+                        android: AndroidChromeCustomTabsOptions(),
+                        ios: IOSSafariOptions(barCollapsingEnabled: true)));
+              },
+              child: Text(
+                tosData.name,
+                style: TextStyle(
+                    decoration: TextDecoration.underline,
+                    color: Theme.of(context).textTheme.button?.color ??
+                        Theme.of(context).primaryColor,
+                    fontSize: 12),
+              ),
+            ),
+          ),
+        if (ppData != null)
+          Container(
+            padding: EdgeInsets.only(left: 2, right: 2),
+            child: GestureDetector(
+              onTap: () async {
+                final ChromeSafariBrowser browser = new ChromeSafariBrowser();
+                await browser.open(
+                    url: Uri.parse(ppData.url),
+                    options: ChromeSafariBrowserClassOptions(
+                        android: AndroidChromeCustomTabsOptions(),
+                        ios: IOSSafariOptions(barCollapsingEnabled: true)));
+              },
+              child: Text(
+                ppData.name,
+                style: TextStyle(
+                    decoration: TextDecoration.underline,
+                    color: Theme.of(context).textTheme.button?.color ??
+                        Theme.of(context).primaryColor,
+                    fontSize: 12),
+              ),
+            ),
+          )
       ],
     );
   }