+++ packages/in_app_purchases_paywall_ui/lib/paywall/simple/simple_paywall_purchase.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:in_app_purchases_intl/helper/paywall_helper.dart';
+import 'package:in_app_purchases_paywall_ui/paywall/inherit/paywall_data_iw.dart';
 import 'package:in_app_purchases_paywall_ui/paywall/inherit/subscription_callback_iw.dart';
-import 'package:in_app_purchases_paywall_ui/paywall/model/icon_and_text.dart';
 import 'package:in_app_purchases_interface/in_app_purchases_interface.dart';
-import 'package:in_app_purchases_paywall_ui/paywall/model/text_and_url.dart';
 import 'package:in_app_purchases_paywall_ui/paywall/simple/legal_row.dart';
 import 'package:in_app_purchases_paywall_ui/paywall/simple/simple_bulletpoints.dart';
 import 'package:in_app_purchases_paywall_ui/paywall/simple/subscription_row.dart';
@@ -10,25 +10,9 @@ import 'package:in_app_purchases_paywall_ui/paywall/simple/subscription_row.dart
 /// This Widget is without a scaffold. Use SimplePayWallScaffold if you want to have an appBar
 ///ignore: must_be_immutable
 class SimplePaywallPurchase extends StatelessWidget {
-  final String? title;
-  final String? subTitle;
-  final TextAndUrl? tosData;
-  final TextAndUrl? ppData;
-  final Widget? headerContainer;
-  final List<IconAndText>? bulletPoints;
-  final Widget? campaignWidget;
-  final String? restoreText;
   bool isSubscriptionLoading = false;
 
   SimplePaywallPurchase({
-    this.title,
-    this.subTitle,
-    this.tosData,
-    this.ppData,
-    this.headerContainer,
-    this.bulletPoints,
-    this.campaignWidget,
-    this.restoreText,
     this.isSubscriptionLoading = false,
     CallbackInterface? callbackInterface,
   });
@@ -50,30 +34,31 @@ class SimplePaywallPurchase extends StatelessWidget {
       Padding(
         padding: EdgeInsets.only(left: 16, top: 32, right: 16, bottom: 16),
         child: Text(
-          title ?? "Go Premium",
+          PaywallDataIW.of(context).title ??
+              PaywallL10NHelper.of(context).app_bar_default_title,
           style: Theme.of(context).textTheme.headline5,
         ),
       ),
     );
 
     // add SubTitle
-    if (subTitle != null) {
+    if (PaywallDataIW.of(context).subTitle != null) {
       elements.add(Container(
           margin: EdgeInsets.only(left: 16, right: 16),
-          child:
-              Text(subTitle!, style: Theme.of(context).textTheme.bodyText2)));
+          child: Text(PaywallDataIW.of(context).subTitle!,
+              style: Theme.of(context).textTheme.bodyText2)));
     }
 
     elements.add(Container(
       padding: EdgeInsets.all(4),
     ));
 
-    if (bulletPoints != null) {
-      elements.add(SimpleBulletPoints(bulletPoints!));
+    if (PaywallDataIW.of(context).bulletPoints != null) {
+      elements.add(SimpleBulletPoints(PaywallDataIW.of(context).bulletPoints!));
     }
 
-    if (campaignWidget != null) {
-      elements.add(campaignWidget!);
+    if (PaywallDataIW.of(context).campaignWidget != null) {
+      elements.add(PaywallDataIW.of(context).campaignWidget!);
     }
 
     elements.add(Container(
@@ -90,7 +75,8 @@ class SimplePaywallPurchase extends StatelessWidget {
       children: [
         TextButton(
           child: Text(
-            restoreText ?? "Restore purchase",
+            PaywallDataIW.of(context).restoreText ??
+                PaywallL10NHelper.of(context).restore_purchase,
           ),
           style: Theme.of(context).textButtonTheme.style,
           onPressed: () {
@@ -103,7 +89,7 @@ class SimplePaywallPurchase extends StatelessWidget {
     elements.add(Container(
       padding: EdgeInsets.all(8),
     ));
-    elements.add(LegalRow(Theme.of(context), tosData, ppData));
+    elements.add(const LegalRow());
 
     elements.add(Container(
       padding: EdgeInsets.all(16),