+++ packages/in_app_purchases_paywall_ui/lib/paywall/widgets/insets/card_insets.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/cupertino.dart';
+
+class CardInsets extends EdgeInsets {
+  CardInsets.normal() : super.only(left: 16, top: 16, right: 16, bottom: 16);
+}